+++ Hobify.Infrastructure/Repository/UserRepository.cs
@@ -45,6 +45,10 @@ namespace Hobify.Infrastructure.Repository
         public async Task<User> DeleteAsync(Guid id)
         {
            var user = await _context.Users.FindAsync(id);
+            if (user == null)
+            {
+                return null;
+            }
             _context.Users.Remove(user);
             await _context.SaveChangesAsync();
             return user;
